refactor: RssFeedAdapter 리펙토링
